namepopup/smaller, less opact markers
Now, there is a pop up when hovering on nodes that shows their names.
*Triangles dont have an exact name but they do have some kind of name that shows first and last leaf of their. So, poped up name and the one we see next to triangle are different.

The markers next to boxplots are a lot smaller and way less opact. But there is no difference when too many markers are in one place.

I can create a pdf raport but the tree image is not full. Working on it.